remove ProtectedRoute component and have AuthContext redirect unauthorised users
